requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pyyaml==6.0.2
markdown==3.7
jinja2==3.1.4
python-multipart==0.0.9
beautifulsoup4==4.12.3
playwright==1.48.0
chromadb>=0.5.0
httpx>=0.27.0
transformers>=4.40.0
pdfplumber>=0.10.0
python-docx>=1.1.0
langchain-text-splitters>=0.2.0
